feat(5): Use Vue Router
Fixes #5
